Remove build system for sip v4
The last bugfix release of sip 4 was in February 2018, 5 years ago. Even
sip 5 is now unsupported. It should be safe to remove the parallel build
system for sip 4 in favor of the sip 5/6 build system defined in
pyproject.toml and project.py, avoiding the need to maintain two build
systems in parallel.